Add validation for empty fields and row selection in Assessment Component form

+++ Forms/Assessment Component.cs
@@ -25,6 +25,14 @@ namespace sample.Forms
             RubricIdBox.DataSource = AssessmentComponentDL.GetRubricIds();
             AssessmentIdBox.DataSource = AssessmentComponentDL.GetAssessmentIds();
         }
+        private bool AreAllFieldEmpty()
+        {
+            if (NameBox.Text == "" && MarksBox.Text == "" && RubricIdBox.SelectedIndex == 0 && AssessmentIdBox.SelectedIndex == 0)
+            {
+                return true;
+            }
+            return false;
+        }
         private void ClearFields()
         {
             NameBox.Text = "";
@@ -34,6 +42,11 @@ namespace sample.Forms
         }
         private void AddButton_Click(object sender, EventArgs e)
         {
+            if (AreAllFieldEmpty())
+            {
+                MessageBox.Show("Please Fill All Fields");
+                return;
+            }
             string name = NameBox.Text;
             int TotalMarks = int.Parse(MarksBox.Text);
             int RubricId = int.Parse(RubricIdBox.SelectedItem.ToString());
@@ -52,6 +65,16 @@ namespace sample.Forms
         }
         private void UpdateButton_Click(object sender, EventArgs e)
         {
+            if (AreAllFieldEmpty())
+            {
+                MessageBox.Show("Please Fill All Fields");
+                return;
+            }
+            if (dataGridView1.CurrentRow == null)
+            {
+                MessageBox.Show("Please Select a Row to Update Data");
+                return;
+            }
             int Id = int.Parse(dataGridView1.CurrentRow.Cells["Id"].Value.ToString());
             string name = NameBox.Text;
             int TotalMarks = int.Parse(MarksBox.Text);
@@ -71,6 +94,12 @@ namespace sample.Forms
         }
         private void DeleteButton_Click(object sender, EventArgs e)
         {
+            // If no row is selected in data grid
+            if (dataGridView1.CurrentRow == null)
+            {
+                MessageBox.Show("Please Select a Row to Delete Data");
+                return;
+            }
             int Id = int.Parse(dataGridView1.CurrentRow.Cells["Id"].Value.ToString());
             if (AssessmentComponentDL.DeleteData(Id))
             {
